Round Two
Had to re-write the constructors for the doctors and the patients. I got the assignment of doctors working, and I'm almost done creating new doctors and patients, but the doctor's specialty and the patient's disease isn't changing between additions. For example if I add two consecutive doctors they will share the same specialty even if I select different ones when constructing them.